+++ R/upgrade.R
@@ -64,9 +64,13 @@ needs_upgrading <- function(object) {
 upgradeParams <- function(params) {
     if (!needs_upgrading(params)) return(params)
     
+    # We'll use the version to decide which upgrades are needed. Copy it to
+    # a variable because params@mizer_version might get changed during the 
+    # upgrade
+    version <- params@mizer_version
+    
     # Before version 2.3 ----
-    if (!.hasSlot(params, "mizer_version") ||
-        params@mizer_version < 2.3) {
+    if (!.hasSlot(params, "mizer_version") || version < 2.3) {
         
         if ("interaction_p" %in% names(params@species_params)) {
             params@species_params$interaction_resource <- 
@@ -300,6 +304,26 @@ upgradeParams <- function(params) {
             pnew@species_params$catch_observed <- NULL
         }
         
+        if (!"Fishing" %in% names(getColours(params))) {
+            params <- setColours(params, c(Fishing = "red"))
+        }
+        if (!"Background" %in% names(getColours(params))) {
+            params <- setColours(params, c(Background = "grey"))
+        }
+        if (!"Total" %in% names(getColours(params))) {
+            params <- setColours(params, c(Total = "black"))
+        }
+        
+        if (!"Fishing" %in% names(getLinetypes(params))) {
+            params <- setLinetypes(params, c(Fishing = "solid"))
+        }
+        if (!"Background" %in% names(getLinetypes(params))) {
+            params <- setLinetypes(params, c(Background = "solid"))
+        }
+        if (!"Total" %in% names(getLinetypes(params))) {
+            params <- setLinetypes(params, c(Total = "solid"))
+        }
+        
         # Copy over all comments
         comment(pnew) <- comment(params)
         for (slot in slotNames(pnew)) {
@@ -312,11 +336,18 @@ upgradeParams <- function(params) {
     }
     
     # Before version 2.4
-    if (params@mizer_version < "2.3.1.9001") {
+    if (version < "2.3.1.9001") {
         # copy w_inf to w_max
         par_names <- names(params@species_params)
         if (!("w_max" %in% par_names)) {
             params@species_params$w_max <- params@species_params$w_inf
+        }
+        # Add linecolour and linetype for external mortality
+        if (!"External" %in% names(getColours(params))) {
+            params <- setColours(params, c(External = "grey"))
+        }
+        if (!"External" %in% names(getLinetypes(params))) {
+            params <- setLinetypes(params, c(External = "solid"))
         }
     }